percentage fix

+++ R8ZAAJ/Model/Food.cs
@@ -8,7 +8,7 @@ namespace R8ZAAJ.Model
         public String Name { get; set; }
         public String FoodName { get; set; }
         public int Price { get; set; }
-        public int EDT { get; set; }
+        public float EDT { get; set; }
         public int LoyaltyPoints { get; set; }
 
         public Food(int id, String name, String foodName, int price, int edt, int loyaltyPoints)
